+++ src/__tests__/advanced.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import App from "../aapp";
+import App from "../app";
 import * as utils from "../utils";
 
 const renderLogMock = vi.spyOn(utils, "renderLog");
