+++ src/components/index.ts
@@ -1,4 +1,4 @@
 export * from "./complex-form";
-export * from "./hheader";
+export * from "./header";
 export * from "./item-list";
 export * from "./notification-system";
