+++ src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import App from "./aapp";
+import App from "./app";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
